+++ tests/__init__.py
@@ -9,9 +9,19 @@ from .integration.test_models.standard_model_tests import (
 from .integration.test_models.standard_output_tests import StandardOutputTests
 from .integration.test_models.standard_output_comparison import StandardOutputComparison
 
-from .unit.test_models.test_full_battery_models.test_lithium_ion.test_lithium_ion_model import (
-    TestLithiumIonModel,
+from .unit.test_models.test_full_battery_models.test_lithium_ion.base_lithium_ion_tests import (
+    BaseUnitTestLithiumIon,
 )
+from .unit.test_models.test_full_battery_models.test_lithium_ion.base_lithium_ion_half_cell_tests import (
+    BaseUnitTestLithiumIonHalfCell,
+)
+from .integration.test_models.test_full_battery_models.test_lithium_ion.base_lithium_ion_tests import (
+    BaseIntegrationTestLithiumIon,
+)
+from .integration.test_models.test_full_battery_models.test_lithium_ion.base_lithium_ion_half_cell_tests import (
+    BaseIntegrationTestLithiumIonHalfCell,
+)
+
 from .shared import (
     get_mesh_for_testing,
     get_p2d_mesh_for_testing,
