+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_dfn_half_cell.py
@@ -1,51 +1,14 @@
 #
 # Tests for the lithium-ion half-cell DFN model
-# This is achieved by using the {"working electrode": "positive"} option
 #
 import pybamm
 import unittest
+from tests import BaseUnitTestLithiumIonHalfCell
 
 
-class TestDFNHalfCell(unittest.TestCase):
-    def test_well_posed(self):
-        options = {"working electrode": "positive"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-
-class TestDFNHalfCellWithSEI(unittest.TestCase):
-    def test_well_posed_constant(self):
-        options = {"working electrode": "positive", "SEI": "constant"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-    def test_well_posed_reaction_limited(self):
-        options = {"working electrode": "positive", "SEI": "reaction limited"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-    def test_well_posed_solvent_diffusion_limited(self):
-        options = {"working electrode": "positive", "SEI": "solvent-diffusion limited"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-    def test_well_posed_electron_migration_limited(self):
-        options = {"working electrode": "positive", "SEI": "electron-migration limited"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-    def test_well_posed_interstitial_diffusion_limited(self):
-        options = {
-            "working electrode": "positive",
-            "SEI": "interstitial-diffusion limited",
-        }
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-    def test_well_posed_ec_reaction_limited(self):
-        options = {"working electrode": "positive", "SEI": "ec reaction limited"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
+class TestDFNHalfCell(BaseUnitTestLithiumIonHalfCell):
+    def setUp(self):
+        self.model = pybamm.lithium_ion.DFN
 
 
 if __name__ == "__main__":
