+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_spme_half_cell.py
@@ -4,53 +4,12 @@
 #
 import pybamm
 import unittest
+from tests import BaseUnitTestLithiumIonHalfCell
 
 
-class TestSPMeHalfCell(unittest.TestCase):
-    def test_well_posed(self):
-        options = {"working electrode": "positive"}
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
-
-    def test_well_posed_integrated(self):
-        options = {"working electrode": "positive"}
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
-
-
-class TestSPMeHalfCellWithSEI(unittest.TestCase):
-    def test_well_posed_constant(self):
-        options = {"working electrode": "positive", "SEI": "constant"}
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
-
-    def test_well_posed_reaction_limited(self):
-        options = {"working electrode": "positive", "SEI": "reaction limited"}
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
-
-    def test_well_posed_solvent_diffusion_limited(self):
-        options = {"working electrode": "positive", "SEI": "solvent-diffusion limited"}
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
-
-    def test_well_posed_electron_migration_limited(self):
-        options = {"working electrode": "positive", "SEI": "electron-migration limited"}
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
-
-    def test_well_posed_interstitial_diffusion_limited(self):
-        options = {
-            "working electrode": "positive",
-            "SEI": "interstitial-diffusion limited",
-        }
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
-
-    def test_well_posed_ec_reaction_limited(self):
-        options = {"working electrode": "positive", "SEI": "ec reaction limited"}
-        model = pybamm.lithium_ion.SPMe(options)
-        model.check_well_posedness()
+class TestSPMeHalfCell(BaseUnitTestLithiumIonHalfCell):
+    def setUp(self):
+        self.model = pybamm.lithium_ion.SPMe
 
 
 if __name__ == "__main__":
